Fixing UUID links
