+++ src/main/java/com/globallogic/amcr/controller/casestudies/CaseStudyController.java
@@ -13,6 +13,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+import org.springframework.web.servlet.view.RedirectView;
 
 import java.util.List;
 import java.util.UUID;
@@ -45,8 +46,8 @@ public class CaseStudyController {
     }
 
     @GetMapping(value = "/{id}/attachment/{attachmentId}")
-    public String getAttachment(@PathVariable UUID id, @PathVariable UUID attachmentId) {
-        return "forward:/attachment/" + attachmentId;
+    public RedirectView getAttachment(@PathVariable UUID id, @PathVariable UUID attachmentId) {
+        return new RedirectView("/attachment/" + attachmentId);
     }
 
     @GetMapping(produces = "application/json")
