+++ src/main/java/com/globallogic/amcr/repository/impl/casestudies/CaseStudyDaoImpl.java
@@ -28,7 +28,7 @@ public class CaseStudyDaoImpl implements CaseStudyDao {
             caseStudyMapper.save(caseStudy);
             return caseStudy;
         } catch (Exception e) {
-            throw new RuntimeException("Error in CaseStudyDaoImpl - could not save case study");
+            throw new RuntimeException("Error in CaseStudyDaoImpl - could not save case study", e);
         }
     }
