+++ src/main/java/com/globallogic/amcr/repository/impl/casestudies/CaseStudyMapper.java
@@ -14,7 +14,7 @@ import java.util.UUID;
 @Mapper
 public interface CaseStudyMapper {
 
-    @Insert("INSERT INTO case_studies(id, spotlight, title, overview, cover_image_id, body, attachment_ids) VALUES (#{id, javaType=java.util.UUID, jdbcType=OTHER, typeHandler=UUIDTypeHandler}, #{spotlight}, #{title}, #{overview}, #{coverImageId, javaType=java.util.UUID, jdbcType=OTHER, typeHandler=UUIDTypeHandler}}, #{body, javaType=java.util.List, jdbcType=OTHER, typeHandler=JSONListMapHandler}, #{attachmentIds, javaType=java.util.Arrays, jdbcType=OTHER, typeHandler=UUIDArrayTypeHandler})")
+    @Insert("INSERT INTO case_studies(id, spotlight, title, overview, cover_image_id, body, attachment_ids) VALUES (#{id, javaType=java.util.UUID, jdbcType=OTHER, typeHandler=UUIDTypeHandler}, #{spotlight}, #{title}, #{overview}, #{coverImageId, javaType=java.util.UUID, jdbcType=OTHER, typeHandler=UUIDTypeHandler}, #{body, javaType=java.util.List, jdbcType=OTHER, typeHandler=JSONListMapHandler}, #{attachmentIds, javaType=java.util.Arrays, jdbcType=OTHER, typeHandler=UUIDArrayTypeHandler})")
     void save(CaseStudy caseStudy);
 
     @Results(id = "caseStudyResults")
@@ -54,7 +54,7 @@ public interface CaseStudyMapper {
     @Select("SELECT id, spotlight, title, overview, cover_image_id FROM case_studies ORDER BY case_study_order DESC LIMIT #{entries}")
     List<CaseStudyOverview> getLatestOverviews(@Param("entries") int entries);
 
-    @Update("UPDATE case_studies SET spotlight = #{caseStudy.spotlight}, title = #{caseStudy.title}, overview = #{caseStudy.overview}, cover_image_id = #{caseStudy.coverImageId, javaType=java.util.UUID, jdbcType=OTHER, typeHandler=UUIDTypeHandler}, body = #{caseStudy.body, javaType=java.util.List, jdbcType=OTHER, typeHandler=JSONListMapHandler}, attachmentIds = #{caseStudy.downloadLinks, javaType=java.util.Arrays, jdbcType=OTHER, typeHandler=UUIDArrayTypeHandler} WHERE id = #{caseStudyId, javaType=java.util.UUID, jdbcType=OTHER, typeHandler=UUIDTypeHandler}")
+    @Update("UPDATE case_studies SET spotlight = #{caseStudy.spotlight}, title = #{caseStudy.title}, overview = #{caseStudy.overview}, cover_image_id = #{caseStudy.coverImageId, javaType=java.util.UUID, jdbcType=OTHER, typeHandler=UUIDTypeHandler}, body = #{caseStudy.body, javaType=java.util.List, jdbcType=OTHER, typeHandler=JSONListMapHandler}, attachment_ids = #{caseStudy.attachmentIds, javaType=java.util.Arrays, jdbcType=OTHER, typeHandler=UUIDArrayTypeHandler} WHERE id = #{caseStudyId, javaType=java.util.UUID, jdbcType=OTHER, typeHandler=UUIDTypeHandler}")
     void update(UUID caseStudyId, CaseStudy caseStudy);
 
     @Delete("DELETE from case_studies WHERE #{id, javaType=java.util.UUID, jdbcType=OTHER, typeHandler=UUIDTypeHandler} = id")
